requirements.txt
pygame>=2.5.0
PyBox2D>=2.3.10
numpy>=1.24.0
